Track more worker-specific stats for all the sites - rtt, status codes, etc.
