+++ lib/httpcheck.js
@@ -74,6 +74,8 @@ var slogger = o_log4js.getLogger( 'slog' );
 var _os         = require( 'os' );
 var hostname    = _os.hostname();
 
+const statsdClient = require( './statsd.js' );
+
 var HttpChecker = {
 	checkServers: function() {
 		try {
@@ -100,13 +102,21 @@ var HttpChecker = {
 		server.processed = true;
 		server.lastCheck = new Date().valueOf();	// we use set the value to the milliseconds value
 
-		if ( rtt > 0 && 400 > http_code && 0 != http_code )
+		if ( rtt > 0 && 400 > http_code && 0 != http_code ) {
 			server.site_status = SITE_RUNNING;
-		else if ( ( SITE_RUNNING == server.oldStatus ) ||
-				( SITE_CONFIRMED_DOWN != server.oldStatus ) && ( new Date().valueOf() < ( server.last_status_change + ( config.get( 'TIME_BETWEEN_NOTICES_MIN' ) * MINUTES ) ) ) )
+		}
+		else if (
+			( SITE_RUNNING == server.oldStatus ) ||
+			(
+				( SITE_CONFIRMED_DOWN != server.oldStatus ) &&
+				( new Date().valueOf() < ( server.last_status_change + ( config.get( 'TIME_BETWEEN_NOTICES_MIN' ) * MINUTES ) ) )
+			)
+		) {
 			server.site_status = SITE_DOWN;
-		else
+		}
+		else {
 			server.site_status = SITE_CONFIRMED_DOWN;
+		}
 
 		if ( server.site_status !=  server.oldStatus ) {
 			var resO    = {};
@@ -160,6 +170,33 @@ var HttpChecker = {
 				running = false;
 			}
 		}
+
+
+		/**
+		 * Log some information in StatsD.
+		 *
+		 * Doing in the end to make sure we send all the data to appropriate consumers first and
+		 * only then we can try to log the data.
+		 */
+		let stats_site_status = '';
+
+		switch ( server.SITE_STATUS ) {
+			case SITE_RUNNING:
+				stats_site_status = 'up';
+				break;
+
+			case SITE_DOWN:
+				stats_site_status = 'down';
+				break;
+
+			case SITE_CONFIRMED_DOWN:
+				stats_site_status = 'still_down';
+				break;
+		}
+
+		statsdClient.increment( `worker.check.${stats_site_status}.code.${http_code}.count` );
+		statsdClient.timing( `worker.check.${stats_site_status}.rtt`, Math.round( rtt / 1000 ) );
+
 	},
 
 	addToQueue: function( arrData ) {
