+++ lib/jetmon.js
@@ -62,7 +62,7 @@ var wpcom    = require( './wpcom'    );
 var comms    = require( './comms'    );
 var cluster  = require( 'cluster'    );
 
-// var http     = require( 'http'       );
+const statsdClient = require('./statsd.js');
 
 var gCountSuccess   = 0;
 var gCountError     = 0;
@@ -80,48 +80,6 @@ var endOfRound      = false;
 global.queuedRetries = [];
 
 logger.debug( 'booting jetmon.js' );
-
-/**
- * Used to get the hostname of the current host, so we can have per-host monitoring of things.
- */
-var _os     = require( 'os' );
-
-/**
- * Hostnames on prod look like:
- * 	<node>.<datacenter>.<domain>
- *
- * We only need the first 2 pieces and flip them around to make easier to group/filter things in StatsD later.
- * @type {string}
- */
-const currentHostname = _os.hostname().split( '.' ).slice( 0, 2 ).reverse().join( '.' );
-
-/**
- * Set up the StatD client.
- *
- * All entries are prefixed with `com.jetpack.jetmon.<hostname>.`
- */
-
-var statsdHostname = '127.0.0.1';
-
-/**
- * Add a workaround for the local Docker instances, as prod is running statsd proxies on 127.0.0.1,
- * while the Docker nodes run it in the `statsd` container.
- */
-if ( currentHostname === 'jetmon.docker' ) {
-	statsdHostname = 'statsd';
-}
-
-var StatsD = require( 'hot-shots' );
-var statsdClient = new StatsD( {
-	host: statsdHostname,
-	port: 8125,
-	globalize: true,
-	prefix: 'com.jetpack.jetmon.' + currentHostname + '.',
-	errorHandler: function( error ) {
-		console.log( "Socket errors caught here: ", error );
-		logger.debug( 'Adding error log' + error );
-	},
-} );
 
 process.on( 'SIGINT',  gracefulShutdown );
 process.on( 'EXIT',    gracefulShutdown );
